implementing the unit of work pattern, adding a confirm dialog service

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
 import { MessageService } from '../_services/message.service';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -16,7 +17,10 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) {}
+  constructor(
+    private messageService: MessageService,
+    private confirmService: ConfirmService
+  ) {}
 
   ngOnInit(): void {
     this.loadMessages();
@@ -36,18 +40,29 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: () => {
-        this.messages?.splice(
-          this.messages.findIndex((m) => m.id === id),
-          1
-        );
-      },
-    });
+    this.confirmService
+      .confirm(
+        'Delete message',
+        'Are you sure you want to delete that message?'
+      )
+      .subscribe((result) => {
+        if (result) {
+          this.messageService.deleteMessage(id).subscribe({
+            next: () => {
+              this.messages?.splice(
+                this.messages.findIndex((m) => m.id === id),
+                1
+              );
+            },
+          });
+        }
+      });
   }
 
   pageChanged(event: any) {
-    this.pageNumber = event.page;
-    this.loadMessages();
+    if (this.pageNumber !== event.page) {
+      this.pageNumber = event.page;
+      this.loadMessages();
+    }
   }
 }
